Added some mutations

+++ gqlapi/database/models/User.go
@@ -1,9 +1,8 @@
 package models
 
-
 // User ...
 type User struct {
-	ID       int64  `json:"id"`
+	ID       int    `json:"id"`
 	Email    string `json:"email"`
 	Username string `json:"username"`
 	Password string `json:"password"`
